+++ src/app/store/UseMovieStore.tsx
@@ -7,89 +7,18 @@ import { toast } from "react-toastify";
 import { Dispatch, RefObject, SetStateAction, createRef } from "react";
 import { TRegisterFormValues } from "../schema/RegisterSchema";
 import { TReviewFormValues } from "../schema/ReviewSchema";
-import { IMovieReviews } from "../interfaces";
-
-export interface IMovie {
-  id: number;
-  name: string;
-  type: string;
-  duration: number;
-  synopsis: string;
-  image: string;
-  reviews: IReview[];
-}
-
-export interface IReview {
-  id: number;
-  movieId: number;
-  userId: number;
-  score: number;
-  description: string;
-}
-
-export interface IUser {
-  email: string;
-  name: string;
-  id: number;
-}
-
-export interface ILogin {
-  accessToken: string;
-  user: IUser;
-}
-
-export interface IMovieStore {
-  movies: undefined | IMovie[];
-  user: undefined | IUser;
-  userLogin: (
-    formData: TLoginFormValues,
-    setLoading: Dispatch<SetStateAction<boolean>>,
-    callback: () => void
-  ) => Promise<void>;
-  userInfo: IUser;
-  autoLogin: (userInfo: IUser) => void;
-  userLogout: (callback: () => void) => void;
-  userRegister: (
-    formData: TRegisterFormValues,
-    setLoading: Dispatch<SetStateAction<boolean>>,
-    callback: () => void
-  ) => Promise<void>;
-  addReviewModalRef: RefObject<HTMLDialogElement>;
-  editReviewModalRef: RefObject<HTMLDialogElement>;
-  deleteReviewModalRef: RefObject<HTMLDialogElement>;
-  addReview: (
-    formData: TReviewFormValues,
-    setLoading: Dispatch<SetStateAction<boolean>>,
-    userId: string,
-    movieId: string,
-    callback: () => void
-  ) => Promise<void>;
-  currentReview: undefined | IReview;
-  fetchReview: (movieId: string, userId: string) => Promise<void>;
-  fetchReviews: (movieId: string) => Promise<void>;
-  deleteReview: (
-    reviewId: string,
-    token: string,
-    setLoading: Dispatch<SetStateAction<boolean>>,
-    callback: () => void
-  ) => Promise<void>;
-  reviews: undefined | IMovieReviews;
-  token: string | null;
-  editReview: (
-    formData: TReviewFormValues,
-    setLoading: Dispatch<SetStateAction<boolean>>,
-    userId: string,
-    movieId: string,
-    reviewId: string,
-    callback: () => void
-  ) => Promise<void>;
-}
+import {
+  ILogin,
+  IMovieReviews,
+  IMovieStore,
+  IReview,
+  IUser,
+} from "../interfaces";
 
 const LOCAL_STORAGE_KEY_USER = "@devmovies_user";
 const LOCAL_STORAGE_KEY_TOKEN = "@devmovies_token";
 
 export const useMovieStore = create<IMovieStore>((set) => ({
-  movies: undefined,
   user: undefined,
   userLogin: async (
     formData: TLoginFormValues,
@@ -151,9 +80,6 @@ export const useMovieStore = create<IMovieStore>((set) => ({
     toast.success("Usuário deslogado com sucesso!");
     callback();
   },
-  addReviewModalRef: createRef<HTMLDialogElement>(),
-  editReviewModalRef: createRef<HTMLDialogElement>(),
-  deleteReviewModalRef: createRef<HTMLDialogElement>(),
   addReview: async (
     formData: TReviewFormValues,
     setLoading: Dispatch<SetStateAction<boolean>>,
@@ -217,34 +143,6 @@ export const useMovieStore = create<IMovieStore>((set) => ({
       setLoading(false);
     }
   },
-  currentReview: undefined,
-  fetchReview: async (movieId: string, userId: string) => {
-    try {
-      const { data } = await api.get<IReview[]>(
-        `/movies/${movieId}/reviews?userId=${String(userId)}`
-      );
-      const filteredReview = await data.filter(
-        (review) => review.id === Number(userId)
-      );
-      set({ currentReview: await filteredReview[0] });
-    } catch (error) {
-      console.log(error);
-    } finally {
-    }
-  },
-
-  reviews: undefined,
-  fetchReviews: async (movieId: string) => {
-    try {
-      const { data } = await api.get<IMovieReviews>(
-        `/movies/${movieId}?_embed=reviews`
-      );
-      console.log(data);
-      set({ reviews: data });
-    } catch (error) {
-      console.log(error);
-    }
-  },
   deleteReview: async (
     reviewId: string,
     token: string,
@@ -265,8 +163,38 @@ export const useMovieStore = create<IMovieStore>((set) => ({
       setLoading(false);
     }
   },
+  fetchReview: async (movieId: string, userId: string) => {
+    try {
+      const { data } = await api.get<IReview[]>(
+        `/movies/${movieId}/reviews?userId=${String(userId)}`
+      );
+      const filteredReview = await data.filter(
+        (review) => review.id === Number(userId)
+      );
+      set({ currentReview: await filteredReview[0] });
+    } catch (error) {
+      console.log(error);
+    } finally {
+    }
+  },
+  currentReview: undefined,
+  reviews: undefined,
+  fetchReviews: async (movieId: string) => {
+    try {
+      const { data } = await api.get<IMovieReviews>(
+        `/movies/${movieId}?_embed=reviews`
+      );
+      console.log(data);
+      set({ reviews: data });
+    } catch (error) {
+      console.log(error);
+    }
+  },
   token:
     typeof window !== "undefined"
       ? localStorage.getItem(LOCAL_STORAGE_KEY_TOKEN)
       : null,
+  addReviewModalRef: createRef<HTMLDialogElement>(),
+  editReviewModalRef: createRef<HTMLDialogElement>(),
+  deleteReviewModalRef: createRef<HTMLDialogElement>(),
 }));
